Scene Transition Testing Phase
Added a new scene with slight changes to emphasize it's a new scene.
Main menu currently on that scene has a start button and will load you into the workshop scene.
The goal is to make a prefab that allows any scene to transition to another one with relative ease.