Save functionality finished for Sprint 3
If objects have been spawned since last save, loading deletes those extra objects. There was a bug where the spawner object was deleted on load if load had to delete extra objects. I fixed that bug.

Still have to handle loading when items have been deleted since save, but I'll do that when we implement ways to remove objects from the scene.

Additionally, the UI doesn't have load/save buttons--it's just F5 to save and F9 to load, and saving currently only creates the SaveData object but does not yet serialize and store it in a file